app/requirements.txt
dotenv==0.0.4
fastapi==0.110.2
passlib==1.7.4
pydantic==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
pyjwt==2.8.0
sqlalchemy==2.0.29
uvicorn==0.29.0